Renamed test files and added tags to them. Removed unused imports.
