+++ src/test/java/uk/gov/companieshouse/accounts/association/integration/UserCompanyAssociationStatusApiIntegrationTest.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.accounts.association.controllers;
+package uk.gov.companieshouse.accounts.association.integration;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -11,7 +11,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.test.web.servlet.MockMvc;
@@ -25,7 +24,7 @@ import uk.gov.companieshouse.accounts.association.repositories.AssociationsRepos
 @SpringBootTest
 @Testcontainers
 @Tag("integration-test")
-class UserCompanyAssociationStatusApiTest {
+class UserCompanyAssociationStatusApiIntegrationTest {
 
     @Container
     @ServiceConnection
